readme added and import cycle error fixed
